# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PortPulse AI API",
    "description": "PortPulse AI: forecasting, anomalies, CP-SAT optimisation, routing, and IBM Bob agent orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plan": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Get Plan",
        "description": "Return the latest persisted plan or build a fresh engine-grounded plan.",
        "operationId": "get_plan_api_plan_get",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Text"
            }
          },
          {
            "name": "persist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Persist"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Regenerate",
        "operationId": "regenerate_api_plan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/vessels/upload": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Upload Schedule",
        "description": "Import a physically complete vessel schedule and immediately rebuild the plan.\n\nPhysical fields are mandatory. The endpoint never substitutes fabricated vessel\ndimensions, draft, moves or destination values.",
        "operationId": "upload_schedule_api_vessels_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_schedule_api_vessels_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ais/import": {
      "post": {
        "tags": [
          "ais"
        ],
        "summary": "Ais Import",
        "description": "Import a real NOAA AccessAIS CSV, aggregate it, and retain filtered track points.",
        "operationId": "ais_import_api_ais_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ais_import_api_ais_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ais/generate": {
      "post": {
        "tags": [
          "ais"
        ],
        "summary": "Ais Generate",
        "description": "Generate synthetic DEMO_AIS data for an offline/reproducible demo.",
        "operationId": "ais_generate_api_ais_generate_post",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 7,
              "default": 14,
              "title": "Days"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20240817,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AisGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AisGenerateResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "rows_scanned": {
            "type": "integer",
            "title": "Rows Scanned"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "vessels": {
            "type": "integer",
            "title": "Vessels"
          },
          "hours": {
            "type": "number",
            "title": "Hours"
          },
          "output_rows": {
            "type": "integer",
            "title": "Output Rows"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "is_measured": {
            "type": "boolean",
            "title": "Is Measured"
          },
          "zones": {
            "type": "integer",
            "title": "Zones"
          },
          "newest_ts": {
            "type": "string",
            "title": "Newest Ts"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "source",
          "message",
          "rows_scanned",
          "skipped",
          "vessels",
          "hours",
          "output_rows",
          "inserted",
          "is_measured",
          "zones",
          "newest_ts",
          "days",
          "seed"
        ],
        "title": "AisGenerateResponse",
        "description": "Keys = build stats + import stats + generator params (all declared)."
      },
...
      "Body_ais_import_api_ais_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ais_import_api_ais_import_post"
      },
...
      "Body_upload_schedule_api_vessels_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_schedule_api_vessels_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanResponse": {
        "properties": {
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "shifts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shifts"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "narrative_source": {
            "type": "string",
            "title": "Narrative Source"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "optimiser_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimiser Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "shifts",
          "text",
          "narrative_source"
        ],
        "title": "PlanResponse",
        "description": "Spec §6 traceability: the plan exposes the snapshot that produced it."
      },
...
      "UploadResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "revisions_created": {
            "type": "integer",
            "title": "Revisions Created"
          },
          "normalised_rows": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalised Rows"
          },
          "upload_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Id"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          },
          "replan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "errors",
          "revisions_created"
        ],
        "title": "UploadResponse",
        "description": "Both upload paths: the parse-failure path omits normalised_rows/replan."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}